Moved models logic
