+++ accdfl/test/core/test_evaluator.py
@@ -2,7 +2,7 @@ import os
 
 import pytest
 
-from accdfl.core.model import create_model
+from accdfl.core.models import create_model
 from accdfl.core.model_evaluator import ModelEvaluator
 from accdfl.core.session_settings import SessionSettings, LearningSettings
