+++ accdfl/test/core/test_model_trainer.py
@@ -2,7 +2,7 @@ import os
 
 import pytest
 
-from accdfl.core.model import create_model
+from accdfl.core.models import create_model
 from accdfl.core.model_trainer import ModelTrainer
 from accdfl.core.session_settings import SessionSettings, LearningSettings
