+++ accdfl/util/evaluate_model.py
@@ -4,7 +4,7 @@ import sys
 
 import torch
 
-from accdfl.core.model import create_model
+from accdfl.core.models import create_model
 from accdfl.core.model_evaluator import ModelEvaluator
 from accdfl.core.session_settings import SessionSettings
