+++ accdfl/util/train_model.py
@@ -4,7 +4,7 @@ import sys
 
 import torch
 
-from accdfl.core.model import create_model
+from accdfl.core.models import create_model
 from accdfl.core.model_trainer import ModelTrainer
 from accdfl.core.session_settings import SessionSettings
