+++ scripts/cifar10.py
@@ -5,7 +5,7 @@ import time
 from accdfl.core.mappings import Linear
 from accdfl.core.model_trainer import ModelTrainer
 from accdfl.core.datasets.CIFAR10 import CIFAR10
-from accdfl.core.model import create_model
+from accdfl.core.models import create_model
 from accdfl.core.session_settings import SessionSettings, LearningSettings
 
 NUM_ROUNDS = 3
