+++ scripts/shakespeare.py
@@ -4,7 +4,7 @@ import time
 
 from accdfl.core.datasets.Shakespeare import Shakespeare
 from accdfl.core.mappings import Linear
-from accdfl.core.model import create_model
+from accdfl.core.models import create_model
 from accdfl.core.model_trainer import ModelTrainer
 from accdfl.core.session_settings import LearningSettings, SessionSettings
